packages/pytest-pydantic-schema-sync/pyproject.toml
[build-system]
requires = [
  "pdm-backend"
]
build-backend = "pdm.backend"

[project]
name = "pytest-pydantic-schema-sync"
version = "0.1.0"
description = "Pytest plugin to synchronise Pydantic model schemas with JSONSchema files"
authors = [
  {name = "Louis Maddox", email = "louismmx@gmail.com"}
]
dependencies = [
  "pydantic-schema-sync>=0.1.5",
  "pytest>=6"
]
requires-python = ">=3.10"
readme = "README.md"

[project.entry-points.pytest11]
pydantic_schema_sync = "pytest_pydantic_schema_sync"

[project.license]
text = "MIT"

[project.urls]
Homepage = "https://github.com/lmmx/pydantic-schema-sync/packages/pytest-pydantic-schema-sync"
Repository = "https://github.com/lmmx/pydantic-schema-sync.git"

[tool.pdm]
distribution = true

[tool.pytest.ini_options.pydantic_schema_sync]
src_dir = "src"
schema_location = "package_root"
schema_dir = "schemas"

[tool.uv.sources]
pydantic-schema-sync = {workspace = true}
